Use load_full until we decide if we use this
