+++ scripts/companydata/requirements.txt
@@ -1,1 +1,1 @@
-psycopg2
+psycopg2-binary
